+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 
 - FEATURE: update textobject to respect `ignore_blank_line` option. Blank lines between commented lines are treated as part of a textobject.
 
+## mini.completion
+
+- FEATURE: respect `isIncomplete` in LSP completion response and immediately force new completion request on the next key press.
+
 ## mini.doc
 
 - FEATURE: improve detection and formatting for types in `@param`, `@return`, and similar.